Move config to testutil.spec.ts
