+++ packages/tendermint-rpc/src/rpcclients/rpcclient.spec.ts
@@ -1,5 +1,5 @@
-import { defaultInstance } from "../config.spec";
 import { createJsonRpcRequest } from "../jsonrpc";
+import { defaultInstance } from "../testutil.spec";
 import { HttpClient } from "./httpclient";
 import { instanceOfRpcStreamingClient } from "./rpcclient";
 import { WebsocketClient } from "./websocketclient";
